Reorganize files: Move shell scripts to scripts/ folder and auth files to auth/ folder
