+++ text.pollinations.ai/auth/googleCloudAuth.js
