+++ text.pollinations.ai/scripts/infinite_loop.sh
