+++ text.pollinations.ai/scripts/setup-google-cloud-auth.sh
