+++ text.pollinations.ai/scripts/test_openai_standard.sh
